+++ xlabuilder/literal_test.go
@@ -17,6 +17,7 @@ func TestLiterals(t *testing.T) {
 	require.NotPanics(t, func() { _ = NewScalarLiteral[complex128](complex(1.0, 0.0)) })
 	require.NotPanics(t, func() { NewScalarLiteral[int8](0).Destroy() })
 	require.NotPanics(t, func() { NewArrayLiteral([]float32{1, 2, 3, 4, 5, 6}, 2, 3).Destroy() })
+	require.NotPanics(t, func() { NewArrayLiteralFromAny([]float64{1, 2, 3, 4, 5, 6}, 2, 3).Destroy() })
 
 	// Check that various literals get correcly interpreted in PRJT.
 	client := getPJRTClient(t)
@@ -34,4 +35,11 @@ func TestLiterals(t *testing.T) {
 	output = capture(Constant(builder, NewScalarLiteralFromAny(float16.Fromfloat32(15e-3)))).Test(t)
 	exec = compile(t, client, capture(builder.Build(output)).Test(t))
 	require.Equal(t, float16.Fromfloat32(15e-3), execScalarOutput[float16.Float16](t, client, exec))
+
+	builder = New(t.Name())
+	output = capture(Constant(builder, NewArrayLiteralFromAny([]float64{1, 3, 5, 7, 11, 13}, 3, 2))).Test(t)
+	exec = compile(t, client, capture(builder.Build(output)).Test(t))
+	gotFlat, gotDims := execArrayOutput[float64](t, client, exec)
+	require.Equal(t, []int{3, 2}, gotDims)
+	require.Equal(t, []float64{1, 3, 5, 7, 11, 13}, gotFlat)
 }