+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@
 * Store client link with Buffer. Added `Buffer.Client` method.
 * Added `Buffer.Device` and `Client.NumForDevice`.
 * Properly setting client options for `pjrt.NewClient`. Added test for reading/writing `C.PJRT_NamedValues`.
+* Added `xlabuilder.Shape.Memory` and `xlabuilder.NewArrayLiteralFromAny`.
 
 # v0.1.2 SuppressAbseilLoggingHack